D-duplicate song and shop fill logic.
These implemetations are logically identical and neither has any song or
shop specific logic. What they do have in common is they both fill the
players own world. So rename the shop implementation to
fill_ownworld_restrictive and delete the copy for songs. No functional
change.

The reason I am doing this, is entrance rando uses the same process as
a third user of this API.