Fix the bubble logic
Fix the bubble logic in child, which needs conditions to determine if
the kill is possible. Revert back to the outdoor regions being adult
only or avoid having to qualify all the internal conditions on those
regions with the reachability criteria which is complicated greately
by the fact that the eye shoot does not stay shot and the SoT block
is absent for child.