Boot disable-state management to World.push_item/pop_item
Get the management of disable location state out of the Fill
implementation and bundle it with push_item.

pop_item() is added to world to be a generic full reversal
of push_item whchi also assume the disable state management.
The fill algorithm will now use pop for its placement failure
rollbacks.