Add support for correlated items and locations
correlated items and locations are items that must be filled as
a set, while each individual item and location still has independant
logic.

The use case is child/adult entrances, for example the child/adult
spirit temple have completely different logic so they need to be
separate, but when rando'ing what gets placed there, you want the
same dungeon to be there for child and adult. This makes the two
different entrances for different ages a correlated location.