Fix age/area playthrough bugs, disable merging.
Wrap sphere area maps and age lists in a Sphere tuple for ease of
tracking. This fixes a couple silly bugs from last night, like not using
the right sphere list.

Unfortunately, something is broken with how the merging algorithm
interacts with this search alg because it breaks the spheres into small
pieces, and sometimes has the wrong ages in place. I am not entirely
sure of the problem, and we may have to rewrite from scratch to fix it.
For now, I'll leave the merge disabled and keep the age/area playthrough.