Make area sphere merging age-aware.
This makes working area playthroughs! With accurate ages!
Still some work to be done, of course, including to add tests to verify
that the age listed in the playthrough is able to access the given
locations.
And figure out what the problem is collecting locations in the final
playthrough step that the fuzzer keeps tripping over...