Change logic so that only the required triforces are logically required
Previously, you would logically require all the triforces, even if not all of them are needed to beat the game. This means that the path of goal hints are once again just way of the hero. The previous path of gold hints are to be replaced with a different hint type.