Add new hint subtype opportunity
This new subtype is a weaker version of woth that is generated when there are optional goals set and there are not enough woth hints. The opportunity hints are like woth, except instead of items to beat the game, then are items needed to complete all the optional objective (and not woth). These are only selected with a 50% chance when there are fewer than 3 woth locations remaining. The random chance is important so that the player cannot know if there are no woth or no opportunity locations remaining. This is the more generalized version of path of gold.